Fix disaster, conflict fields and move sync logic to task
