+++ apps/gidd/schema.py
@@ -27,7 +27,7 @@ from apps.country.models import Country
 def disaster_statistics_qs(disaster_qs) -> DisasterStatisticsType:
     timeseries_qs = disaster_qs.filter(new_displacement__gt=0).values('year').annotate(
         total=Coalesce(Sum('new_displacement', output_field=IntegerField()), 0)
-    ).order_by('year').values('year', 'total')
+    ).order_by('year').values('year', 'total', 'country_id', 'country_name', 'iso3')
 
     # FIXME should we filter out not labeld hazard type?
     categories_qs = disaster_qs.filter(hazard_type__isnull=False).values('hazard_type').annotate(
@@ -50,6 +50,11 @@ def disaster_statistics_qs(disaster_qs) -> DisasterStatisticsType:
         timeseries=[DisasterTimeSeriesStatisticsType(
             year=item['year'],
             total=item['total'],
+            country=DisasterCountryType(
+                id=item['country_id'],
+                iso3=item['iso3'],
+                country_name=item['country_name']
+            )
         ) for item in timeseries_qs],
 
         categories=[CategoryStatisticsType(**item) for item in categories_qs]
